Mar-7-2019_4:38_PM

+++ de.bat
@@ -114,7 +114,7 @@ rem lu: Jan-2-2019
 echo.
 echo %fp%
 
-del /s "*conflicted copy*.*"
+if exist *conflicted copy*.* del /s "*conflicted copy*.*"
 
 exit/b
